requirements.txt
allure-pytest==2.11.0
allure-python-commons==2.11.0
iniconfig==1.1.1
pytest==7.1.3
pytest-order==1.0.1
selenium==4.5.0
undetected-chromedriver==3.1.5.post4
